requirements.txt
# requirements
snowflake-snowpark-python
streamlit
requests
pandas